Split cli.py into a cli/ package
