+++ koffee/cli/__main__.py
@@ -1,0 +1,6 @@
+"""Entry point for `python -m koffee.cli`."""
+
+from koffee.cli.commands import main
+
+if __name__ == "__main__":
+    main()
