+++ koffee/cli/commands.py
@@ -1,27 +1,21 @@
-"""The koffee CLI."""
+"""CLI command implementations."""
 
 import logging
 import shutil
 import subprocess
-import tomllib
 from pathlib import Path
 from typing import Annotated
 
-from cyclopts import App, Group, Parameter, validators
-from pydantic import ValidationError
+from cyclopts import Parameter, validators
 from rich.console import Console
-from rich.logging import RichHandler
-from rich.progress import (
-    BarColumn,
-    Progress,
-    SpinnerColumn,
-    TextColumn,
-    TimeElapsedColumn,
-)
+from rich.progress import Progress
 from rich.table import Table
 
 from koffee import asr
 from koffee.api import SUBTITLE_EXTENSIONS, SUPPORTED_EXTENSIONS, _write_output, run
+from koffee.cli.app import app, defaults, log, options_group
+from koffee.cli.embedded import _handle_embedded_subtitles
+from koffee.cli.progress import _create_progress_bar, _make_progress_callback
 from koffee.embed import embed_subtitles
 from koffee.exceptions import KoffeeError
 from koffee.schemas.config import (
@@ -33,58 +27,6 @@ from koffee.schemas.config import (
 )
 from koffee.subtitle import generate_subtitles
 from koffee.utils import get_subtitle_tracks, parse_subtitle_file
-
-logging.basicConfig(
-    level=logging.INFO, format="%(message)s", datefmt="[%X]", handlers=[RichHandler()]
-)
-
-log = logging.getLogger(__name__)
-
-app = App(
-    default_parameter=Parameter(negative=""),
-    group_arguments=Group("Arguments", sort_key=0),
-    group_commands=Group("Commands", sort_key=1),
-    group_parameters=Group("Parameters", sort_key=2),
-    name="koffee",
-    version_flags=["--version", "-V"],
-)
-
-options_group = Group("Options", sort_key=3)
-
-app["--help"].group = options_group
-app["--version"].group = options_group
-
-
-def _load_cli_defaults() -> KoffeeConfig:
-    """Loads CLI defaults from the config file, falling back on invalid input."""
-    try:
-        return KoffeeConfig(**load_config_file())
-    except (ValidationError, tomllib.TOMLDecodeError) as exc:
-        log.warning(f"Ignoring invalid config file for CLI defaults: {exc}")
-        return KoffeeConfig()
-
-
-defaults = _load_cli_defaults()
-
-
-def _create_progress_bar() -> Progress:
-    """Creates a rich progress bar for tracking transcription and translation."""
-    return Progress(
-        SpinnerColumn(),
-        TextColumn("[bold blue]{task.description}"),
-        BarColumn(),
-        TextColumn("{task.percentage:>3.0f}%"),
-        TimeElapsedColumn(),
-    )
-
-
-def _make_progress_callback(progress: Progress, task_id) -> callable:
-    """Returns a callback that updates a progress bar given a 0.0-1.0 ratio."""
-
-    def callback(ratio: float) -> None:
-        progress.update(task_id, completed=ratio * 100)
-
-    return callback
 
 
 @app.default()
@@ -260,67 +202,6 @@ def _print_dry_run(resolved_paths: list[Path], config: KoffeeConfig) -> None:
         log.info(f"[dry-run] Subtitles will be embedded into video ({config.embed})")
 
 
-def _handle_embedded_subtitles(video_path: Path, config: KoffeeConfig) -> KoffeeConfig:
-    """If the video has embedded subtitles, prompts the user and updates config."""
-    tracks = _detect_embedded_subtitles(video_path)
-    if not tracks:
-        return config
-
-    log.info(f"Found {len(tracks)} embedded subtitle track(s) in {video_path.name}.")
-    if not _prompt_use_embedded_subtitles():
-        return config
-
-    return _apply_subtitle_track(config, tracks)
-
-
-def _apply_subtitle_track(config: KoffeeConfig, tracks: list[dict]) -> KoffeeConfig:
-    """Selects a subtitle track and returns an updated config."""
-    track_index, source_language = _select_subtitle_track(tracks)
-    updates = {"use_embedded_subtitles": True, "subtitle_track_index": track_index}
-    if source_language:
-        updates["source_language"] = source_language
-
-    return config.model_copy(update=updates)
-
-
-def _detect_embedded_subtitles(video_path: Path) -> list[dict]:
-    """Returns embedded subtitle tracks in the video, or an empty list."""
-    if video_path.suffix.lower() in SUBTITLE_EXTENSIONS:
-        return []
-
-    return get_subtitle_tracks(video_path)
-
-
-def _prompt_use_embedded_subtitles() -> bool:
-    """Prompts the user to use embedded subtitles instead of running ASR."""
-    user_input = input("Translate embedded subtitles instead of running ASR? [Y/n] ")
-    return user_input.strip().lower() in ("", "y", "yes")
-
-
-def _select_subtitle_track(tracks: list[dict]) -> tuple[int, str | None]:
-    """Prompts user to select a subtitle track if multiple are available."""
-    if len(tracks) == 1:
-        language = tracks[0].get("tags", {}).get("language")
-        return 0, language
-
-    log.info("Available subtitle tracks:")
-    for i, track in enumerate(tracks):
-        tags = track.get("tags", {})
-        language = tags.get("language", "unknown")
-        title = tags.get("title", "")
-        label = f"  [{i}] {language}"
-        if title:
-            label += f" — {title}"
-        log.info(label)
-
-    user_input = input(f"Select track [0-{len(tracks) - 1}] (default 0): ")
-    index = int(user_input.strip()) if user_input.strip().isdigit() else 0
-    index = max(0, min(index, len(tracks) - 1))
-
-    language = tracks[index].get("tags", {}).get("language")
-    return index, language
-
-
 def _resolve_paths(file_path: tuple) -> list[Path]:
     """Resolves glob patterns, directories, and files into a flat list of paths."""
     resolved_paths = []
@@ -393,93 +274,43 @@ def _translate_with_progress(
 
 
 @app.command()
-def info() -> None:
-    """Display system information for debugging."""
-    log.info("[koffee info]")
-
-    ffmpeg_path = shutil.which("ffmpeg")
-    if ffmpeg_path:
-        result = subprocess.run(
-            ["ffmpeg", "-version"],
-            check=False,
-            capture_output=True,
-            text=True,
-            timeout=10,
-        )
-        version_line = result.stdout.split("\n")[0]
-        log.info(f"  ffmpeg: {version_line}")
-    else:
-        log.info("  ffmpeg: not found")
-
-    ffprobe_path = shutil.which("ffprobe")
-    log.info(f"  ffprobe: {'found' if ffprobe_path else 'not found'}")
-
-    try:
-        import torch  # noqa: PLC0415
-
-        cuda_available = torch.cuda.is_available()
-        device_name = torch.cuda.get_device_name(0) if cuda_available else "N/A"
-        log.info(f"  CUDA: {'available' if cuda_available else 'not available'}")
-        if cuda_available:
-            log.info(f"  GPU: {device_name}")
-    except ImportError:
-        log.info("  torch: not installed")
-
-    config = KoffeeConfig(**load_config_file())
-    log.info(f"  default whisper model: {config.whisper_model}")
-    log.info(f"  default backend: {config.provider}")
-    log.info(f"  config file: {_find_config_path() or 'none'}")
-
-
-@app.command()
-def languages() -> None:
-    """List all supported language codes."""
-    num_columns = 4
-    codes = sorted(LANGUAGE_CODES - {"auto"})
-    entries = [f"{code} ({LANGUAGE_NAMES.get(code, code)})" for code in codes]
-
-    table = Table(show_header=False, box=None, pad_edge=False, expand=True)
-    for _ in range(num_columns):
-        table.add_column()
-
-    for i in range(0, len(entries), num_columns):
-        row = entries[i : i + num_columns]
-        row += [""] * (num_columns - len(row))
-        table.add_row(*row)
-
-    console = Console()
-    console.print(table)
-    console.print(f"\n[bold]{len(codes)}[/bold] supported languages")
-
-
-def _find_config_path() -> Path | None:
-    """Returns the path to the active config file, or None."""
-    for path in CONFIG_SEARCH_PATHS:
-        if path.is_file():
-            return path
-    return None
-
-
-@app.command()
-def tracks(
+def convert(
     file_path: Annotated[Path, Parameter(validator=validators.Path(exists=True))],
+    subtitle_format: Annotated[str, Parameter(name=("--format", "-f"))] = "vtt",
+    output_dir: Annotated[Path, Parameter(name=("--output-dir", "-o"))] | None = None,
+    output_name: Annotated[str, Parameter(name=("--output-name", "-n"))] | None = None,
+    overwrite: Annotated[
+        bool, Parameter(name=("--overwrite",), group=options_group)
+    ] = False,
 ) -> None:
-    """List embedded subtitle tracks in a video file."""
-    track_list = get_subtitle_tracks(file_path)
+    """Convert a subtitle file between formats (SRT, VTT, ASS).
 
-    if not track_list:
-        log.info(f"No subtitle tracks found in {file_path.name}.")
-        return
+    Parameters
+    ----------
+    file_path: Path
+        Path to the subtitle file
+    subtitle_format: str
+        Target subtitle format (srt, vtt, or ass)
+    output_dir: Path
+        Directory for the output file
+    output_name: str
+        Name of the output file
+    overwrite: bool
+        Overwrite existing output files instead of raising an error
+    """
+    segments = parse_subtitle_file(file_path)
+    out_dir = output_dir if output_dir is not None else file_path.parent
+    subtitle_path = generate_subtitles(subtitle_format, segments, out_dir)
 
-    log.info(f"Subtitle tracks in {file_path.name}:")
-    for i, track in enumerate(track_list):
-        tags = track.get("tags", {})
-        language = tags.get("language", "unknown")
-        title = tags.get("title", "")
-        label = f"  [{i}] {language}"
-        if title:
-            label += f" — {title}"
-        log.info(label)
+    target_path = _write_output(
+        subtitle_path,
+        file_path,
+        subtitle_format,
+        output_dir,
+        output_name,
+        overwrite,
+    )
+    log.info(f"Converted {file_path.name} to {target_path}")
 
 
 @app.command()
@@ -518,6 +349,96 @@ def embed(
 
     result = embed_subtitles(subtitle_path, video_path, output_path, mode=mode)
     log.info(f"Output saved to {result}")
+
+
+@app.command()
+def info() -> None:
+    """Display system information for debugging."""
+    log.info("[koffee info]")
+
+    ffmpeg_path = shutil.which("ffmpeg")
+    if ffmpeg_path:
+        result = subprocess.run(
+            ["ffmpeg", "-version"],
+            check=False,
+            capture_output=True,
+            text=True,
+            timeout=10,
+        )
+        version_line = result.stdout.split("\n")[0]
+        log.info(f"  ffmpeg: {version_line}")
+    else:
+        log.info("  ffmpeg: not found")
+
+    ffprobe_path = shutil.which("ffprobe")
+    log.info(f"  ffprobe: {'found' if ffprobe_path else 'not found'}")
+
+    try:
+        import torch  # noqa: PLC0415
+
+        cuda_available = torch.cuda.is_available()
+        device_name = torch.cuda.get_device_name(0) if cuda_available else "N/A"
+        log.info(f"  CUDA: {'available' if cuda_available else 'not available'}")
+        if cuda_available:
+            log.info(f"  GPU: {device_name}")
+    except ImportError:
+        log.info("  torch: not installed")
+
+    config = KoffeeConfig(**load_config_file())
+    log.info(f"  default whisper model: {config.whisper_model}")
+    log.info(f"  default backend: {config.provider}")
+    log.info(f"  config file: {_find_config_path() or 'none'}")
+
+
+def _find_config_path() -> Path | None:
+    """Returns the path to the active config file, or None."""
+    for path in CONFIG_SEARCH_PATHS:
+        if path.is_file():
+            return path
+    return None
+
+
+@app.command()
+def languages() -> None:
+    """List all supported language codes."""
+    num_columns = 4
+    codes = sorted(LANGUAGE_CODES - {"auto"})
+    entries = [f"{code} ({LANGUAGE_NAMES.get(code, code)})" for code in codes]
+
+    table = Table(show_header=False, box=None, pad_edge=False, expand=True)
+    for _ in range(num_columns):
+        table.add_column()
+
+    for i in range(0, len(entries), num_columns):
+        row = entries[i : i + num_columns]
+        row += [""] * (num_columns - len(row))
+        table.add_row(*row)
+
+    console = Console()
+    console.print(table)
+    console.print(f"\n[bold]{len(codes)}[/bold] supported languages")
+
+
+@app.command()
+def tracks(
+    file_path: Annotated[Path, Parameter(validator=validators.Path(exists=True))],
+) -> None:
+    """List embedded subtitle tracks in a video file."""
+    track_list = get_subtitle_tracks(file_path)
+
+    if not track_list:
+        log.info(f"No subtitle tracks found in {file_path.name}.")
+        return
+
+    log.info(f"Subtitle tracks in {file_path.name}:")
+    for i, track in enumerate(track_list):
+        tags = track.get("tags", {})
+        language = tags.get("language", "unknown")
+        title = tags.get("title", "")
+        label = f"  [{i}] {language}"
+        if title:
+            label += f" — {title}"
+        log.info(label)
 
 
 @app.command()
@@ -594,46 +515,6 @@ def transcribe(
     log.info(f"Output saved to {target_path}")
 
 
-@app.command()
-def convert(
-    file_path: Annotated[Path, Parameter(validator=validators.Path(exists=True))],
-    subtitle_format: Annotated[str, Parameter(name=("--format", "-f"))] = "vtt",
-    output_dir: Annotated[Path, Parameter(name=("--output-dir", "-o"))] | None = None,
-    output_name: Annotated[str, Parameter(name=("--output-name", "-n"))] | None = None,
-    overwrite: Annotated[
-        bool, Parameter(name=("--overwrite",), group=options_group)
-    ] = False,
-) -> None:
-    """Convert a subtitle file between formats (SRT, VTT, ASS).
-
-    Parameters
-    ----------
-    file_path: Path
-        Path to the subtitle file
-    subtitle_format: str
-        Target subtitle format (srt, vtt, or ass)
-    output_dir: Path
-        Directory for the output file
-    output_name: str
-        Name of the output file
-    overwrite: bool
-        Overwrite existing output files instead of raising an error
-    """
-    segments = parse_subtitle_file(file_path)
-    out_dir = output_dir if output_dir is not None else file_path.parent
-    subtitle_path = generate_subtitles(subtitle_format, segments, out_dir)
-
-    target_path = _write_output(
-        subtitle_path,
-        file_path,
-        subtitle_format,
-        output_dir,
-        output_name,
-        overwrite,
-    )
-    log.info(f"Converted {file_path.name} to {target_path}")
-
-
 app["info"].sort_key = 0
 app["languages"].sort_key = 1
 app["tracks"].sort_key = 2
@@ -645,7 +526,3 @@ app["embed"].sort_key = 5
 def main() -> None:
     """Wraps app() so that it is accessible to poetry."""
     app()
-
-
-if __name__ == "__main__":
-    main()
